Unused code removed, Added comments & bifurcation of files
